+++ src/lib/helpers/url-params.jsx
@@ -3,6 +3,14 @@ import { console_debug_log } from "../services/logging.service.jsx";
 const debug = false;
 
 export function getUrlParams(props = window) {
+    /*
+    Get query parameters / url parameters
+
+    Example:
+        const getUrlParams = gs.urlParams.getUrlParams;
+        const urlParams = getUrlParams();
+        noMenu = (urlParams.menu && urlParams.menu === "0"),
+    */
     let urlParams = {};
     let searchString;
     if (debug) console_debug_log('getUrlParams | props:', props);
@@ -22,7 +30,7 @@ export function getUrlParams(props = window) {
                     return urlParams;
                 }
                 let keyPairs = searchString.split("&");
-                if(Array.isArray(keyPairs)) {
+                if (Array.isArray(keyPairs)) {
                     for (let i = 0; i < keyPairs.length; i++) {
                         const keyPairString = keyPairs[i];
                         const [rawKey, ...rest] = keyPairString.split('=');
@@ -53,7 +61,7 @@ export function getUrlParams(props = window) {
                     urlParams = props.match.params;
                 }
             }
-        }   
+        }
     } catch (error) {
         console.log(`getUrlParams ERROR | ${props}`);
         console.error(error)