+++ src/lib/services/generic.editor.rfc.selector.jsx
@@ -143,6 +143,18 @@ export const GenericSelectGenerator = (props) => {
 };
 
 export const GenericSelectDataPopulator = (props) => {
+  /*
+   * Generic select data populator component.
+   * Return the data for a select, sending a request to the API, and adding a <option>...</option> with the key and description for each row returned
+   *
+   * Parameters:
+   *  filter: filter by _id. Default to no filter (null)
+   *  dbFilter: database query filter. Default to no filter (null)
+   *  columns: columns to show in the listing page or read-only form page. Default is "" meaning all columns
+   *  title_field_name: field name to show in the title. Default is "title"
+   *  value_field_name: field name to show in the value. Default is "value"
+   *  key_name: field name to show in the key. Default is "_id"
+   */
   const [errorState, setErrorState] = useState(null);
   const [config, setConfig] = useState(null);
   const [rows, setRows] = useState(null);